robotics-book/chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.3
qdrant-client==1.11.3
psycopg[binary]==3.2.3
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
python-multipart==0.0.12
markdown==3.7
beautifulsoup4==4.12.3
tiktoken==0.8.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
